learn/mcp/requirements.txt
mcp>=1.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
langgraph>=0.2.0
langchain>=0.3.0
